+++ service/controller/key/cluster.go
@@ -73,7 +73,8 @@ const (
 )
 
 const (
-	LifeCycleHookNodePool = "NodePool"
+	LifeCycleHookControlPlane = "ControlPlane"
+	LifeCycleHookNodePool     = "NodePool"
 )
 
 const (