+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Control Plane drainer controller.
+
 ### Fixed
 
 - Fix aws operator policy for latest node pools version.
+- Make encryption key lookup graceful during cluster creation.
+
 
 
 ## [8.2.3] 2020-04-06
@@ -30,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use architect orb.
 - Fix subnet allocation for Availability Zones.
 - Switch to AWS CNI
+
+
 
 ## [8.2.1] - 2020-03-20